[core] Do the MapContext::cleanup() synchronously
This used to be async assuming the libuv behavior of holding the
main loop until all the messages are processed and also for avoiding
a deadlock at exit that no longer exits.

Now that we support multiple main loops implementation on different
platforms, we do the cleanup synchronously to make sure all the
platforms have the same behavior at exit.

Fixes #4631